backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pycryptodome==3.19.0
opencv-python-headless==4.7.0.72
numpy==1.25.2
pillow==10.0.1
tqdm==4.65.0
reedsolo==1.6.1
soundfile==0.12.1
requests==2.31.0
PyJWT==2.10.1
Werkzeug==3.1.4
cryptography==46.0.3
argon2-cffi==25.1.0
cffi==2.0.0
